Fix: cardService 로직 분리 #59
