+++ src/main/java/WELLET/welletServer/card/service/CardService.java
@@ -103,36 +103,4 @@ public class CardService {
 
         return new CardCountResponseDto((long) cards.size(), cards);
     }
-
-    @Transactional
-    public Card saveCard (Long memberId, SaveMyCard dto) {
-        cardRepository.findByOwnerId(memberId).ifPresent(e -> {
-            throw new CardException(CardErrorCode.DUPLICATE_MY_CARD);
-        });
-
-        Card card = Card.builder()
-                .name(dto.getName())
-                .position(dto.getPosition())
-                .email(dto.getEmail())
-                .phone(dto.getPhone())
-                .tel(dto.getTel())
-                .department(dto.getDepartment())
-                .company(dto.getCompany())
-                .address(dto.getAddress())
-                .ownerId(memberId)
-                .build();
-
-        return cardRepository.save(card);
-    }
-
-    public CardResponse findMyCard(Long memberId) {
-        Card card = cardRepository.findByOwnerId(memberId)
-                .orElse(null);
-
-        if (card == null) {
-            return null;
-        }
-
-        return CardResponse.toCardDto(card);
-    }
 }
